editing added

+++ GiacintDllExpo/Lib/Services/TextEditor.cs
@@ -6,35 +6,58 @@ using System.Runtime.InteropServices;
 namespace GiacintDllExpo.Lib.Services;
 internal class TextEditor
 {
+    private static bool isOpened = false;
+
     internal async static Task<string> Edit(string text)
     {
-        string tempFile = Path.Combine(Path.GetTempPath(), $"temp_edit_{Guid.NewGuid()}.txt");
-        File.WriteAllText(tempFile, text);
-
-        ProcessStartInfo psi;
-
-        if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-            psi = new ProcessStartInfo("notepad.exe", $"\"{tempFile}\"");
-
-        else if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
-            psi = new ProcessStartInfo("xdg-open", $"\"{tempFile}\"");
-        else if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
-            psi = new ProcessStartInfo("open", $"\"{tempFile}\"");
-        else
-            throw new PlatformNotSupportedException("Unsupported OS");
-
-        psi.UseShellExecute = true;
-
-        _ = Task.Run(() =>
+        if (isOpened)
         {
-            using (var proc = Process.Start(psi))
-            {
-                proc.WaitForExit();
-            }
-        });
+            Debug.Warning("Text editor is already opened, returning original text.");
+            return text;
+        }
 
-        string result = await File.ReadAllTextAsync(tempFile);
-        File.Delete(tempFile);
-        return result;
+        isOpened = true;
+        try
+        {
+            string tempFile = Path.Combine(Path.GetTempPath(), $"temp_edit_{Guid.NewGuid()}.txt");
+            File.WriteAllText(tempFile, text);
+
+            ProcessStartInfo psi;
+
+            if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
+                psi = new ProcessStartInfo("notepad.exe", $"\"{tempFile}\"");
+
+            else if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
+                psi = new ProcessStartInfo("xdg-open", $"\"{tempFile}\"");
+            else if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
+                psi = new ProcessStartInfo("open", $"\"{tempFile}\"");
+            else
+                throw new PlatformNotSupportedException("Unsupported OS");
+
+            psi.UseShellExecute = true;
+
+            _ = Task.Run(() =>
+            {
+                using (var proc = Process.Start(psi))
+                {
+                    proc.WaitForExit();
+                }
+            });
+
+            string result = await File.ReadAllTextAsync(tempFile);
+            File.Delete(tempFile);
+
+            isOpened = false;
+            return result;
+        }
+        catch (Exception ex)
+        {
+            Debug.Error(ex);
+            return text;
+        }
+        finally
+        {
+            isOpened = false;
+        }
     }
 }
